Added in the Stiffeners and the Shearplates
Created the methods for adding the shear plates and the stiffeners
and added them in with the profile
When 9/16 holes become studs then the studs are on both flanges now
when the beam is above the
- fixed a bug with the arc direction not defaulting to the inside or
outside of the arc based on the Z vector of the direction being zero.
now the defaulting is predictable.
- also updated the wide flange to have the right steel laering on the
bolts